fix: merge refactoring
